moved code to main
